dcp_step_table added
added a table where pre-calculated step values are stored. helps simplify determining the correct number of steps to take on a given value of dsubmode in the color picker. color picker is nearly there. thinking about 'white' as its own, single value, in order to gain a larger range of colors